src/services/keyword-research/requirements.txt
flask==2.0.1
nltk==3.6.3
scikit-learn==0.24.2
pandas==1.3.3
numpy==1.21.2
requests==2.26.0
python-dotenv==0.19.0
gunicorn==20.1.0
prometheus-client==0.11.0
pytest==6.2.5
